+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - changed `<r>` key to `<Ctrl-r>` and `<d>` key to `<Ctrl-d>`
 - added `<r>` key (and `random` argument) to toggle between random/sequential mode
+- allow adjusting the line width during play with the `<(>` and `<)>` keys
 
 # chesstrainer 0.1-0 (2024-12-17)